Public API

BLASBenchmarksCPU.benchmark_result_df

BLASBenchmarksCPU.benchmark_result_threaded

BLASBenchmarksCPU.benchmark_result_type

BLASBenchmarksCPU.logspace

BLASBenchmarksCPU.plot

BLASBenchmarksCPU.runbench

BLASBenchmarksCPU.benchmark_result_df — Method

benchmark_result_df(benchmark_result::BenchmarkResult)

BLASBenchmarksCPU.benchmark_result_threaded — Method

benchmark_result_threaded(benchmark_result::BenchmarkResult)

BLASBenchmarksCPU.benchmark_result_type — Method

benchmark_result_type(benchmark_result::BenchmarkResult)

BLASBenchmarksCPU.logspace — Method

logspace(start, stop, length)

Defines a monotonically increasing range, log spaced when possible. Useful for defining a range of sizes for benchmarks.

julia> collect(logspace(1,100,3))
3-element Vector{Int64}:
   1
  10
 100

julia> collect(logspace(1,10,3))
3-element Vector{Int64}:
  1
  3
 10

julia> collect(logspace(1,5,3))
3-element Vector{Int64}:
 1
 2
 5

julia> collect(logspace(1,3,3))
3-element Vector{Int64}:
 1
 2
 3

BLASBenchmarksCPU.plot — Method

plot(br::BenchmarkResult;
     desc = "",
     logscale = true,
     width = 1200,
     height = 600,
     plot_directory = default_plot_directory(),
     plot_filename = default_plot_filename(br; desc = desc, logscale = logscale),
     file_extensions = ["svg", "png"])

BLASBenchmarksCPU.runbench — Method

runbench(T = Float64;
         libs = default_libs(T),
         sizes = logspace(2, 4000, 200),
         threaded::Bool = Threads.nthreads() > 1,
         A_transform = identity,
         B_transform = identity,
         sleep_time = 0.0)